functionOC/requirements.txt
Azure-functions
implicit==0.5.2
scipy
